@@ -5,13 +5,13 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.7.15</version>
+        <version>3.2.1</version>
         <relativePath/>
     </parent>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <snow.version>2.0.1-SNAPSHOT</snow.version>
         <snow.util.version>2.0.1-SNAPSHOT</snow.util.version>
         <snow.common.version>2.0.1-SNAPSHOT</snow.common.version>
