@@ -20,7 +20,7 @@
 	<name>snow-eureka</name>
 	<description>snow-eureka</description>
 	<properties>
-		<java.version>1.8</java.version>
+		<java.version>17</java.version>
 		<spring-cloud.version>2021.0.6</spring-cloud.version>
 	</properties>
 	<dependencies>
