@@ -1,6 +1,6 @@
 package com.stylesmile.modules.system.entity;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 
 /**
  * @author chenye
