@@ -1,51 +1,51 @@
-package com.stylesmile.common.configure;
-
-
-import com.alibaba.druid.support.http.StatViewServlet;
-import com.alibaba.druid.support.http.WebStatFilter;
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-import org.springframework.boot.web.servlet.ServletRegistrationBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * Created by chen on 2017/5/16.
- * <p>
- * <p>
- * Describe:
- */
-@Configuration
-public class DruidConfig {
-    /**
-     * 注册DruidServlet
-     * http://localhost:8080/druid/datasource.html查看监控信息
-     * @return
-     */
-    @Bean
-    public ServletRegistrationBean druidServletRegistrationBean() {
-        ServletRegistrationBean servletRegistrationBean = new ServletRegistrationBean();
-        servletRegistrationBean.setServlet(new StatViewServlet());
-        servletRegistrationBean.addUrlMappings("/druid/*");
-        return servletRegistrationBean;
-    }
-
-    /**
-     * 注册DruidFilter拦截
-     *
-     * @return
-     */
-    @Bean
-    public FilterRegistrationBean duridFilterRegistrationBean() {
-        FilterRegistrationBean filterRegistrationBean = new FilterRegistrationBean();
-        filterRegistrationBean.setFilter(new WebStatFilter());
-        Map<String, String> initParams = new HashMap<String, String>();
-        //设置忽略请求
-        initParams.put("exclusions", "*.js,*.gif,*.jpg,*.bmp,*.png,*.css,*.ico,/druid/*");
-        filterRegistrationBean.setInitParameters(initParams);
-        filterRegistrationBean.addUrlPatterns("/*");
-        return filterRegistrationBean;
-    }
-}
+//package com.stylesmile.common.configure;
+//
+//
+//import com.alibaba.druid.support.http.StatViewServlet;
+//import com.alibaba.druid.support.http.WebStatFilter;
+//import org.springframework.boot.web.servlet.FilterRegistrationBean;
+//import org.springframework.boot.web.servlet.ServletRegistrationBean;
+//import org.springframework.context.annotation.Bean;
+//import org.springframework.context.annotation.Configuration;
+//
+//import java.util.HashMap;
+//import java.util.Map;
+//
+///**
+// * Created by chen on 2017/5/16.
+// * <p>
+// * <p>
+// * Describe:
+// */
+//@Configuration
+//public class DruidConfig {
+//    /**
+//     * 注册DruidServlet
+//     * http://localhost:8080/druid/datasource.html查看监控信息
+//     * @return
+//     */
+//    @Bean
+//    public ServletRegistrationBean druidServletRegistrationBean() {
+//        ServletRegistrationBean servletRegistrationBean = new ServletRegistrationBean();
+//        servletRegistrationBean.setServlet(new StatViewServlet());
+//        servletRegistrationBean.addUrlMappings("/druid/*");
+//        return servletRegistrationBean;
+//    }
+//
+//    /**
+//     * 注册DruidFilter拦截
+//     *
+//     * @return
+//     */
+//    @Bean
+//    public FilterRegistrationBean duridFilterRegistrationBean() {
+//        FilterRegistrationBean filterRegistrationBean = new FilterRegistrationBean();
+//        filterRegistrationBean.setFilter(new WebStatFilter());
+//        Map<String, String> initParams = new HashMap<String, String>();
+//        //设置忽略请求
+//        initParams.put("exclusions", "*.js,*.gif,*.jpg,*.bmp,*.png,*.css,*.ico,/druid/*");
+//        filterRegistrationBean.setInitParameters(initParams);
+//        filterRegistrationBean.addUrlPatterns("/*");
+//        return filterRegistrationBean;
+//    }
+//}
