@@ -1,8 +1,7 @@
 package com.stylesmile.modules.system.entity;
 
+import jakarta.validation.constraints.NotEmpty;
 import org.hibernate.validator.constraints.Length;
-
-import javax.validation.constraints.NotEmpty;
 
 /**
  * @author chenye
